+++ doc/release_notes.rst
@@ -2,16 +2,26 @@
 specter Release Notes
 =====================
 
-0.5.0 (unreleased)
+0.6.1 (unreleased)
+------------------
+
+* no changes yet
+
+0.6.0 (2016-08-16)
+------------------
+
+PR #40:
+* Added full_output option to ex2d to get model image and metrics based upon
+  goodness of fit
+* PSFs can specify their model error with PSFERR header keyword; default 0.01
+
+0.5.0 (2016-05-23)
 ------------------
 
 * Move data files into Python package so pip can install the data files.
 * Load test files in class methods to hopefully speed up tests.
 * Improve Travis test support to latest standards.
 * Added a documentation page for the specter API.
-* Added full_output option to ex2d to get model image and metrics based upon
-  goodness of fit
-* PSFs can specify their model error with PSFERR header keyword; default 0.01
 
 0.4.1 (2016-03-10)
 ------------------